+++ src/trkbs_api/__init__.py
@@ -6,9 +6,11 @@ from .headings import (
     tag_heading,
     tag_heading_stream,
 )
-from .tag_sub import (
+from .edit_tags import (
     replace_tag,
     replace_tag_stream,
+    replace_tag_name,
+    replace_tag_name_stream,
     replace_attr,
     replace_attr_stream,
 )
@@ -17,8 +19,15 @@ from .regesta import (
     add_regesta,
 )
 from .marginalia import tag_marginalia, tag_marginalia_stream
-from .export import get_text, format_page_xml
-from .validate import count_tag_by_page
+from .export_text import get_text, format_page_xml
+from .count_structure import count_tag_by_page
+from .search_tags import (
+    TagInstance,
+    find_tags,
+    find_tags_by_page,
+    count_tags,
+    count_tags_by_page,
+)
 
 __all__ = [
     "TrkbsClient",
@@ -29,6 +38,8 @@ __all__ = [
     "tag_heading_stream",
     "replace_tag",
     "replace_tag_stream",
+    "replace_tag_name",
+    "replace_tag_name_stream",
     "replace_attr",
     "replace_attr_stream",
     "tag_empty_lines",
@@ -38,4 +49,9 @@ __all__ = [
     "get_text",
     "format_page_xml",
     "count_tag_by_page",
+    "TagInstance",
+    "find_tags",
+    "find_tags_by_page",
+    "count_tags",
+    "count_tags_by_page",
 ]